Address code review feedback
- Add null check for empty response in CourtImagesService
- Remove unnecessary cast for _court.Id (already int type)

+++ APP.Eds/APP.Eds/UsesCases/Court/CourtDetailPage.xaml.cs
@@ -187,7 +187,7 @@ public partial class CourtDetailPage : ContentPage
             ImagesLoadingIndicator.IsRunning = true;
 
             var imagesService = new CourtImagesService();
-            var images = await imagesService.GetCourtImagesAsync((int)_court.Id);
+            var images = await imagesService.GetCourtImagesAsync(_court.Id);
 
             CourtImages.Clear();
             if (images != null && images.Any())
